+++ CHANGELOG.md
@@ -43,11 +43,15 @@ Ref: https://keepachangelog.com/en/1.0.0/
 
 * (log) [#948](https://github.com/tharsis/ethermint/pull/948) redirect go-ethereum's logs to Cosmos SDK logger.
 
+### Features
+* (ante) [#950](https://github.com/tharsis/ethermint/pull/950) Add support for EIP712 signed Cosmos transactions
+
 ### Bug Fixes
 
 * (rpc) [#955](https://github.com/tharsis/ethermint/pull/955) Fix websocket server push duplicated messages to subscriber.
 * (rpc) [tharsis#953](https://github.com/tharsis/ethermint/pull/953) Add `eth_signTypedData` api support.
 * (log) [#948](https://github.com/tharsis/ethermint/pull/948) redirect go-ethereum's logs to cosmos-sdk logger.
+
 
 ## [v0.10.0-beta1] - 2022-02-15